+++ lib/live_select.ex
@@ -18,8 +18,8 @@ defmodule LiveSelect do
   clicking an option with the mouse.
 
   Whenever an option is selected, `LiveSelect` will trigger a standard `phx-change` event in the form. See the "Examples" section
-  below for details on how to handle the event. In `single` and `tags` mode, the input text and the list of selectable options is cleared after selection.
-  To preserve both, use the `keep_options_on_select` flag in the assigns.
+  below for details on how to handle the event. In `single` and `tags` mode, the content of the input text field and the list of selectable options are cleared on selection.
+  To suppress this behavior, use the `keep_options_on_select` flag in the assigns.
 
   In single mode, if the configuration option `allow_clear` is set, the user can manually clear the selection by clicking on the `x` button on the input field.
   In tags mode, single tags can be removed by clicking on them.
@@ -416,7 +416,7 @@ defmodule LiveSelect do
 
   attr :keep_options_on_select, :boolean,
     doc:
-      ~s(if `true`, when in single or tags mode, the current set of options and the input text is preserved upon selection)
+      ~s(if `true`, the current list of selectable options and the content of the input text field are preserved upon selection)
 
   attr :value, :any, doc: "used to manually set a selection - overrides any values from the form.
   Must be a single element in `:single` mode, or a list of elements in `:tags` mode."