+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 ## [Unreleased]
 
 ### Added
-* DaisyUI 5 compatibility - active options now include both `active` and `menu-active` classes for compatibility with DaisyUI 3, 4, and 5.
-* Add `keep_options_on_select` toggle
+* DaisyUI 5 compatibility - active options now include both `active` and `menu-active` classes for compatibility with DaisyUI 3, 4, and 5
+* Add `keep_options_on_select` flag. By default, when not in `quick_tags` mode, input text field and list of selectable options are now cleared after selection. This flag suppresses this behavior   
 * Styling clear buttons entirely using JS
 
 ## 1.6.0 (2025-04-13)